Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/tee.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL tee.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
